feat: cleanup and avoid circular deps
